+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Notable changes to this project are documented in this file.
 
-# Unreleased
+# 1.7.0
 
 ## Fixed
 - do not use update channel for Toolbox managed apps